backend/requirements.txt
# backend
flask
flask-cors

#ai
langchain
langchain-google-genai
#google

#env
python-dotenv

#database
supabase

#docs
pypdf